Update return type of `HuffLZ::GetNextCode` to avoid type conversion
Fix Clang warning `-Wimplicit-int-conversion`.

Fix MSVC warning:
> warning C4244: '=': conversion from 'int' to 'unsigned short', possible loss of data

Error was diagnosed at line 167, where `GetNextCode()` was called.

+++ src/Archive/HuffLZ.cpp
@@ -195,7 +195,7 @@ namespace OP2Utility::Archive
 
 
 	// Performs a tree lookup guided by the bitstream to find the next code
-	int HuffLZ::GetNextCode()
+	unsigned short HuffLZ::GetNextCode()
 	{
 		int nodeIndex;
 		bool bBit;
